tests/submission/W3C/domhtml/HTMLLabelElement04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - LABEL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-label-element.html#dom-label-htmlfor" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post">
<p>
<label accesskey="b" for="input1">Enter Your First Password:</label>
<input id="input1" name="Password1" type="password">
</p>
</form>
<p>
<label accesskey="c" for="input2">Enter Your Second Password:</label>
<input id="input2" name="Password2" type="password">
</p>




<script>
var domtest = async_test("The htmlFor attribute links this label with another form control by     id attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhtmlfor;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("label");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(0);
      vhtmlfor = testNode.htmlFor;

      assert_equals(vhtmlfor,"input1");
       
    domtest.done();
}));

</script>
<p>The htmlFor attribute links this label with another form control by
    id attribute.

    Retrieve the htmlFor attribute of the first LABEL element 
    and examine its value.</p>
<div id="log"></div>
</body></html>
